enhancement: Compound expressions without parentheses
[CEL does not short-circuit when evaluating logical
expressions](https://github.com/google/cel-spec/blob/master/doc/langdef.md#logical-operators).
Because of this, if a condition contains a compound logical expression,
it should be parenthesised to obtain the desired result.

For example:

```
match:
  all:
    of:
      - expr: a
      - expr: b || c
```

should be written in one of the following ways to be correct:

```
match:
  all:
    of:
      - expr: a
      - expr: (b || c)
```

```
match:
  all:
    of:
      - expr: a
      - any:
         of:
           - expr: b
           - expr: c
```

This PR automatically parenthesises expressions so that users do not
have to remember to do it themselves.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>