docs: Document the query plan response
Describes the format of the query plan response with a list of possible
operators and some examples of ASTs.

The CEL-specific operators such as `loop`, `loop-accu-init` etc. have
been omitted from the list for the moment but they should be documented
with examples in a future PR.

These docs assume that the changes made in #558 have been merged.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>