enhancement: Allow aux_data to be accessed as auxData (#623)
This is an inconsistency stemming from the way the CEL library handles
protobuf field names. It's easy for users to be confused because in the
request the field is named `auxData` but in the CEL expression it can
only be accessed as `aux_data`. This patch fixes that inconsistency.

Fixes #504

Signed-off-by: Charith Ellawala <charith@cerbos.dev>