chore(e2e): Add E2E test with tracing enabled (#907)
The issue in #905 wasn't caught early because tracing is initialised in
`main` that is not covered by unit tests. The E2E tests didn't catch it
either because tracing was disabled. This PR adds an E2E test that runs
with Jaeger tracing enabled so that we can catch similar issues in the
future.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>