chore(test): Update Minio config in tests
The blob store tests have started failing because the new release of
Minio container is checking the data directory for IO-DIRECT access.
Switching to the default `/data` directory seems to fix this issue.

In addition, the healthcheck should be using the readiness probe instead
of the liveness probe because sometimes the server is not yet ready to
accept connections even if it is up.

Also reduces the parallelism in E2E test runs because lately they have
been failing -- possibly due to resource contention.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>