fix: Treat non-boolean expression results as `false`
When evaluating `any`, `all` or `none` blocks, if one of the expressions
returns a non-boolean result, the whole block is marked as having
failed. We can be a bit more lenient here and treat non-boolean values
as `false`. Note that this treats _all_ values as `false` and does not
make the distinction of treating certain values as `true`. Users are
responsible for converting any non-boolean values into boolean values
themselves.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>