chore: Stable string representation for query filter
When generating the string representation of the query plan filter, we
use the `protojson` package to marshal some of the CEL protobufs.
Unfortunately, the output of `protojson` is not guaranteed to be stable
and the library intentionally adds random whitespaces to the output.

This PR strips non-essential whitespace from the JSON bytes to make the
output usable in tests.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>