fix: Handle other data types while optimizing IN expression
When the query planner is trying to optimize an IN expression, it is
wrongly assuming that the operand can only be a list. This leads to the
planner completely ignoring IN expressions that don't have a list as an
operand.

This fix changes to logic to handle both lists and maps, which are valid
operand types to `IN`. Other non-container types cause the `IN` operator
to automatically convert itself to an `EQ`.

Fixes #1333

Signed-off-by: Charith Ellawala <charith@cerbos.dev>