fix: Handle principal-policy-only cases
If there's only a principal policy and at least one of the actions does
not have a matching rule in the principal policy, everything gets denied
because we don't have a resource policy to fall back to. This PR fixes
that issue and preserves the valid decisions that came out of the
principal policy.

Fixes #1397

Signed-off-by: Charith Ellawala <charith@cerbos.dev>