fix: Evict policies that are changed in-place
If a user edits a policy file in place and changes its identifiers
(name, version) then the old policy does not get evicted from the
compile cache and is still usable until Cerbos is restarted or the store
is reloaded using the Admin API.

This PR also adds a configuration field to set the expiry time of
compiled policies. The default is no expiry, which is the current
behaviour.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>