enhancement!: Make config flag optional
Make the `--config` flag optional for launching Cerbos. If it's not
specified and `CERBOS_CONFIG` environment variable is not set as well,
a default configuration is applied which sets the policy directory to
`${PWD}/policies`.

This is a breaking change for users who are using a container
orchestrator that uses Docker healthchecks (e.g. ECS) AND has mounted a
custom configuration file at `/conf.default.yaml` because the
container's `CERBOS_CONFIG` environment variable no longer points to it.
They should manually set the `CERBOS_CONFIG` environment variable as a
workaround. Unfortunately it's difficult to keep the old behaviour
without either duplicating the default configuration in two places or
breaking the `healthcheck` command.

Includes a couple of changes to fix deprecated settings in GoReleaser as
well.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>