enhancement: Add support for reading archive files for the disk driver (#1473)
Allow users to specify zip, tar and gzip archive files for disk storage drivers.

The contents of the archive is expected to be the same as a standard policy directory; non-policy yaml files cannot be included. The archive is targeted by specifying it's relative location as the "directory" in the config. Directory watch is automatically disabled when specifying an archive file.

Signed-off-by: Sam Lock <sam@swlock.co.uk>