chore(test): Remove manual setup from overlay E2E test (#1633)
* chore(test): Remove manual setup from overlay E2E test

Avoid having to manually generate the database script to bootstrap the
overlay test.

This is a bit of a hack. It starts a second Cerbos instance connected to
the same Postgres instance and runs the bootstrapping tests to load the
policies into the database. Then, by reusing the same `*testing.T`, it
prevents the resources from being torn down and runs the second set of
tests on the same database.

Fixes #1612

Signed-off-by: Charith Ellawala <charith@cerbos.dev>

* Delete the SQL script

Signed-off-by: Charith Ellawala <charith@cerbos.dev>

---------

Signed-off-by: Charith Ellawala <charith@cerbos.dev>