feat: Lenient scope search
When configured with `--set=engine.lenientScopeSearch=true`, performs a
search down the scope tree until a policy is found. Originally, if the
request defined scope `a.b.c` a policy file with scope `a.b.c` had to
exist in the store. With this configuration if `a.b.c` doesn't exist,
the engine will try to load `a.b`, `a` and `` in that order.

Fixes #1587

Signed-off-by: Charith Ellawala <charith@cerbos.dev>