enhancement: Clean-up store resources (#1749)
* enhancement: Clean-up store resources

The new version of tar-fs does not have to read the entire archive into
memory. However, it requires the tar file handle to stay open for the
lifetime of the FS. Because the index did not hold closable resources
until now, this PR adds the capability to clean-up resources held by the
index. It also fixes a few places where resources held by stores were not
getting cleaned up properly.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>

* Fix lint warnings

Signed-off-by: Charith Ellawala <charith@cerbos.dev>

---------

Signed-off-by: Charith Ellawala <charith@cerbos.dev>