enhancement(api): Separate Go module for API definitions
Make the generated protobuf code available as a separate module so that
they can be shared with other projects without pulling in all Cerbos
dependencies.

The `replace` directive in `go.mod` will have to be removed in a
separate PR because it's impossible to work with unpublished modules in
Go without `replace` directives.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>