Use toolchain to manage Go version used
By specifying `go 1.21.1` in the module any consumers of this module will be forced to use `1.21.1` as a minimum version.

Now that the generated protobuf contracts are their own module, trying to depend on them forced us to use `1.21.1`.

This change introduces `toolchain` which means that a developer or the build of the main Cerbos project will automatically download this version to use (it's a module), while allowing anyone with `>=1.21.0` to still use the module.

Signed-off-by: Rob Crowe <282571+rcrowe@users.noreply.github.com>
Signed-off-by: Rob Crowe <nobby.crowe@gmail.com>