chore: Upgrade to CEL 0.18
This is the minimum required to work with CEL 0.18.x in the short term.
Further work is required to use the native Go types instead of the
expression protobufs because that's the direction the CEL library is
heading. However, that requires a significant refactoring of the code
base.

Part of #1812

Signed-off-by: Charith Ellawala <charith@cerbos.dev>