feat: Better support for OTLP  (#1886)
Add ability to configure OTLP trace exporters with more options such as
the choice of protocols, sampler configuration, TLS settings etc.

As part of this change, the `tracing` configuration block in the Cerbos
configuration file has been completely deprecated with the aim of
removing it in the release after next. This is because of the following
reasons:
- Jaeger native protocol is no longer supported by the Otel SDK.
- The Otel specification defines standard environment variables that can
be used to configure OTLP exporters. Trying to replicate all possible
configuration options in our configuration would be brittle and just
complicate our code and documentation for not much benefit.

Fixes #1784 
Part of #341

---------

Signed-off-by: Charith Ellawala <charith@cerbos.dev>