fix: Restore `User-Agent` header aliasing (#1941)
It turns out that gRPC-Gateway overwrites the forwarded `User-Agent`
header with something like `grpc-gateway grpc-go/1.60.1`. This restores
the behaviour prior to #1934 where the incoming `User-Agent` is
forwarded as `Grpcgateway-User-Agent`.

Also, this reduces the minimum flush interval on the local audit log
backend to 1 second, which aligns with the confdoc example values. Doing
so makes it easier to query the audit logs shortly after sending a
request to test that the SDKs are sending metadata correctly.

---------

Signed-off-by: Andrew Haines <haines@cerbos.dev>