fix: Record HTTP remote address as peer address for HTTP requests
The peer address stored in the audit log is the gRPC peer address. When
there's a HTTP request, because it goes through the gRPC gateway, the
peer address reported by gRPC metadata is the localhost address.

This PR updates the gateway configuration to capture the HTTP remote
address and store it in metadata so that the audit log can record the
actual originator of the request.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>