chore: Add ability to parse well-known types
Adds support for parsing protobuf WKTs which require special handling.
Because the WKT handling code is private, it cannot be used directly
from our code. Since we only use WKTs in test suite and fixture
definitions, they are offloaded to the `protojson` package for parsing.

Also includes a fix to the quoted string detector so that it doesn't
produce false positives for flow style objects (JSON embedded in YAML
such as `foo: {"bar": "baz"}`).

Signed-off-by: Charith Ellawala <charith@cerbos.dev>