feat: Initial support for cerboshub audit log backend (#1999)
This adds initial support for the `cerboshub` audit log backend. It
extends the existing `local` backend with functionality to sync logs
with Cerbos Hub (RPC implementation to follow). Separately namespaced
keys are added with audit or decision logs, which are subsequently
scanned, paginated, flushed and deleted (atomically) on an interval.

This is a WIP -- there are a few outstanding items to address in future
PRs:

- Integration with `svc-ingest`
- Handle generation of confdocs for embedded type
- Log filtering

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>