docs(backport): Add recipe for permission checks in the UI (#2038)
We get asked how to do permissions checks in the UI fairly frequently
(e.g.
[here](https://community.cerbos.dev/t/8043690/hi-all-is-there-any-way-api-to-get-all-allowed-resources-and#a908848c-9a30-45be-a28d-9d7c94d1c9ca),
[here](https://community.cerbos.dev/t/12647108/hi-wave-is-there-a-best-practice-for-conditional-rendering-o#ef19573b-c759-4ec1-bc0a-cbb42cdf0268),
[here](https://github.com/cerbos/cerbos/discussions/1603), and
[here](https://community.cerbos.dev/t/16650698/hi-guys-i-have-a-stupid-question-slightly-smiling-face-i-und#163d8d5c-c84c-4cc1-bc28-2311d3772840)),
so it'd be good to provide some recommended approaches.

I also replaced some of the uses of "Title Case" with "Sentence case"
because the docs currently use a bit of both, which is like nails on
chalkboard to me 😆

---------

Signed-off-by: Andrew Haines <haines@cerbos.dev>