fix: Skip compile cache if first candidate doesn't exist
When lenient scope search is enabled and we have a set of candidates to
search for, we incorrectly search the cache for all of those and return
the first one found. If we don't find the first candidate in the cache,
we should just hit the store because it could have been added to the
repo since we last looked for it.

Fixes #1846

Signed-off-by: Charith Ellawala <charith@cerbos.dev>