docs: Add OpenAPI documentation
This adds a statically generated OpenAPI reference to the docs site.
Currently it opens in a new window and doesn't link back to the main
docs site. While it's possible to embed the component into the docs
site, the navigation links don't work in that mode.

Having the static docs available on the site makes it possible to create
links to certain RPCs when answering questions. Therefore, it's still
useful to have them even if they are not natively integrated.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>