chore: Remove callback from local audit log
The callback is in the write path and blocks until the Hub log is done
with its work using its internal timer. By removing this dependency, we
can let the two logs work independently. Even if the Hub sync falls
behind and adds delays to the internal timer, the local log saving
thread can continue to work without being blocked.

The down side of this change is that tests are a bit more flaky.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>