feat: Add Role policies (#2192) (#2260)
Role policies are simple RBAC policies in which you specify a number
of resources; each with a set of permissible actions that the role can
carry out on the resource. You can view it as a per-role matrix of
resource/action rows/columns whereby each set "cell" is an
`ALLOW`. They do not support conditions.

During evaluation of a check request, a role policy is only evaluated
if it has a matching principal role and scope. If a principal has more
than one role, the evaluator can return more than one role policy.
The union of the `resource` lists is then taken which provides a
larger potential pool of matching `resource:action` pairs.

Policies are evaluated from most to least specific. If any role policies
are matched according to the above rule, they are evaluated before
any resource policies, but after any principal policies.

If a `resource:action` pairing is not explicitly set in any of the role
policies, that pairing will immediately resolve to a `DENY`.
Otherwise, that given `resource:action` pairing will fall through to
any resource policies for further evaluation.

Signed-off-by: Sam Lock <sam@swlock.co.uk>