backport(docs): Update Helm doc (#2278) (#2279)
PAT should be the password as per
https://docs.cerbos.dev/cerbos/latest/configuration/storage#git-driver

#### Description

<!-- Thank you for contributing Cerbos! Please describe the changes made
in this PR here and provide any other useful information for reviewers.
Make sure that you included some automated tests (e.g unit tests) to
verify your changes. If there is a requirement for user input for
testing, please include the instructions as well. -->

Fixes #<!-- Link the relevant issue here -->

#### Checklist 

<!-- See
https://github.com/cerbos/cerbos/blob/main/CONTRIBUTING.md#submitting-pull-requests
for more information. -->

- [ ] The PR title has the correct prefix 
- [ ] PR is linked to the corresponding issue
- [ ] All commits are signed-off (`git commit -s ...`) to provide the
[DCO](https://developercertificate.org/)

Signed-off-by: Alex Olivier <alex@alexolivier.me>