enhancement: Remove eager log initialisation from schema validation (#2287)
We construct a logger with the incoming payload and pass it to the
validate function where it's only ever used if validation fails. For
large payloads, this adds a noticeable cost. Furthermore, because the
error message already includes object paths, it's not necessary to log
the original payload as well.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>

Signed-off-by: Charith Ellawala <charith@cerbos.dev>