enhancement: Introduce policy scope FALL_THROUGH_ON_ALLOW strategy (#2325)
This PR introduces a policy-agnostic attribute `ScopePermissions` with
two options:

- `SCOPE_PERMISSIONS_OVERRIDE_PARENT`: this is pre-existing behaviour
for resource and principal policies. If the scoped policy results in
allow or explicit deny, it returns that; otherwise (if the scoped policy
results in no match) it falls through to the parent.
- `SCOPE_PERMISSIONS_REQUIRE_PARENTAL_CONSENT_FOR_ALLOWS`: this is
pre-existing behaviour for newly introduced role policies. If the scoped
policy results in allow, it falls through to be evaluated by the parent
scopes. For no match, role policies return an immediate deny, whereas
resource and principal policies will fall through to the parent for further
evaluation. All policy types return an immediate deny for explicit denies.

Any policy type is now free to implement either strategy. If
unimplemented, it will default to existing behaviour.

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>