enhancement: Role policies parent roles field
Introduce a discrete set of parent roles which a role policy:
* implicitly assumes
* narrows the permissions of

If a principal has a custom role, and the role policy definition for
that role has one or more parent roles, the principal implicitly assumes
those custom roles.

It also allows role policies to narrow the permissions of more than one
IdP role (rather than just the named role, as was previously the case).

Signed-off-by: Sam Lock <sam@swlock.co.uk>