chore(release): Update generated API module version before tagging (#2350)
Because we have a `go.work` file, the version of `api/genpb` submodule
in `go.mod` is ignored in workspace-aware builds. This is generally what
we want because it makes it much easier to test changes to the
protobufs. However, when we tag a release, we should make sure that the
`go.mod` points to the actual version of the submodule used. This is
important for reproducible builds because Go workspaces are considered a
development aid by some tools and they disable workspace mode when
building from source.

I am currently having trouble building Cerbos 0.39 with Nix because the
tagged version uses v0.38.1 of `api/genpb` which doesn't contain some of
the protobuf definitions used in 0.39.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>

Signed-off-by: Charith Ellawala <charith@cerbos.dev>