feat: Use groups of principals and/or resources in test cases (#2340)
This PR introduces `principalGroups` and `resourceGroups`, which may be
defined in test fixture files or in test suites and then referenced in
the inputs and expectations of test cases.

I've relaxed the rule about specifying multiple expectations for the
same principal+resource pair because in the case of groups with
overlapping members you might want to specify expectations for these
separately. Now, we merge the expectations and only throw an error if
they are inconsistent.

A group defined in a test suite overrides one with the same key in the
fixtures file, the same as fixtures themselves. A group in a fixtures
file can only refer to fixtures defined within that file. A group
defined in a test suite may refer to shared fixtures and/or fixtures
defined in the suite.

Signed-off-by: Andrew Haines <haines@cerbos.dev>