enhancement: Make `--verbose` less noisy in combination with `--run` (#2351)
If you have a large test suite and you run into failures that you want
to debug with traces, running that specific test with `--run` and
specifying `--verbose` gives you the traces, but also gives you a
massive tree showing all the tests that were skipped because they didn't
match the pattern you provided.

Taking a leaf out of `go test -run`'s book, this PR hides the tests that
were skipped because they didn't match the pattern.

As a bonus, this means that the `skipReason` set on test suites or
individual tests is now propagated to the test results, which means that
we can start displaying it in Hub.

I've also fixed an error in the logic for filtering tests when policies
have been excluded from a bundle.

---------

Signed-off-by: Andrew Haines <haines@cerbos.dev>