enhancement: Role policies parent roles field (#2349)
Introduce a set of parent roles which a role policy assumes.

* Parent roles can originate from role policies defined from the policy
set. Therefore, at role policy set retrieval time, we recursively
retrieve and aggregate all possible role policies.
* Both `scopePermissions` settings are honoured. "Override parent"
inherits the permissions of parent roles and has the authority to return
`ALLOW` early if a `resource:action` pair exists in any of the policies.
"Required parental consent for allows" narrows the scope for the full
set of (recursive) parent roles.
* Role policies with "override parent" now return `NO_MATCH` (rather
than `DENY`) for `resource:action` pairs that don't exist. This covers
the case where we need to inherit a base (non role policy) role, and
evaluate it in the resource policies in the case where a
`resource:action` pair isn't defined in the role policy. It also brings
the behaviour in line with resource policy types.
* This change also introduces parent role support into the query
planner.

There is scope for some caching improvements to the recursive lookup,
but I don't think this is super straight forward as cache invalidation
requires context from a separate area of concern within the PDP (we'd
need to be aware of index updates from within the engine). This is
doable, but we don't know if it's necessary yet, and we're going to
revamp the index at some point anyway so that might come first.

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>