docs: Update what-is-cerbos.adoc tenant ->tenet (#2406)
tenet : a principle, belief, or doctrine generally held to be true
especially : one held in common by members of an organization, movement,
or profession

tenant : one who has the occupation or temporary possession of lands or
tenements of another
specifically : one who rents or leases a dwelling (such as a house) from
a landlord

#### Description

<!-- Thank you for contributing Cerbos! Please describe the changes made
in this PR here and provide any other useful information for reviewers.
Make sure that you included some automated tests (e.g unit tests) to
verify your changes. If there is a requirement for user input for
testing, please include the instructions as well. -->

Fixes #<!-- Link the relevant issue here -->

#### Checklist 

<!-- See
https://github.com/cerbos/cerbos/blob/main/CONTRIBUTING.md#submitting-pull-requests
for more information. -->

- [ ] The PR title has the correct prefix 
- [ ] PR is linked to the corresponding issue
- [ ] All commits are signed-off (`git commit -s ...`) to provide the
[DCO](https://developercertificate.org/)

Signed-off-by: John P. Costello M.Sc. <johnpc123@gmail.com>