chore: Protect against non leaf REQUIRE_PARENTAL_CONSENT scopes (#2493)
The documentation states that any `REQUIRE_PARENTAL_CONSENT` scopes must
reside at a leaf node in the scope hierarchy, but we do not enforce it
in the PDP.

We're not far off supporting this (middle layer `*_CONSENT`), but adding
tests uncovered some issues (namely around parent-role row retrieval).
Separately, I'm not sure if it would even be a worthwhile feature.

Therefore, adding build time protection seems viable for now.

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>