chore: Decouple role policies from scope permissions (#2496)
Role policies are now agnostic to scope permissions. They can reside
inside any scope and exhibit the same consistent behaviour: they inherit
and narrow their parent roles.

If a parent role points to another role policy, then it can only ALLOW
the maximum set of permissions specified in that parent role. This is
evaluated recursively.

If a parent is an IdP role, this IdP role can have rules present in the
same scope or inside a parent scope. Those policies follow the standard
scope permissions based evaluation patterns.

I've also removed the restriction on scopePermission/scope ordering so
scopes with `scopePermissions: REQUIRE_PARENTAL_CONSENT` are no longer
constrained to the leaf nodes. Conversely, a scope with
`OVERRIDE_PARENT` can now theoretically be a descendent of an
`REQUIRE_PARENTAL_CONSENT` scope (although I'm unaware of a reasonable
use-case for this).

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>