fix: Tidy up rule table trace outputs (#2531)
This tidies up the trace output so it's more sensical and more in line
with previous behaviour.

I opted to _not_ implement a resource trace output as it is inferred by
the FQN, and I removed the role trace output as this was misleading.
There might be value in separating out trace outputs to distinguish
between resource and role policies, but this is tricky and would require
some reworking (role policies are just pass throughs that generate
un-distinguishing rows at present).

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>