feat(plan): Add support for multiple actions (#2543)
Add support for multiple actions in the query plan API. The produced
query plan is a logical AND of the individual query plans for each
action. It can be used to obtain a list of resources on which a
principal can perform all the listed actions.

In other words, the combined query plan represents the intersection of
permissions across all requested actions.

---------

Signed-off-by: Dennis Buduev <dbuduev@users.noreply.github.com>
Signed-off-by: dbuduev <dennis@cerbos.dev>
Co-authored-by: Dennis Buduev <dbuduev@users.noreply.github.com>
Co-authored-by: Charith Ellawala <charithe@users.noreply.github.com>