feat: Decouple rule table (#2617)
This is the first part of an extended effort to provide a dedicated
package for rule table management and consumption (primarily to avoid a
single ginormous PR).

The focus of this first change is to:
- Separate the rule table logic into it's own dedicated package
- To introduce an object that accepts precompiled rule tables and
provides structured APIs for `Check` and `Plan` requests.

The following items will be tackled in followup PRs:
- Elegant handling of policy store update events (perhaps via a rule
table patching mechanism) to avoid full lazy refreshes.
- Further simplification of the `Check` and `Plan` methods, by:
- Removing the dependency on the PolicyLoader--the ruletable `Manager`
should only take static, precompiled rule tables as input and should not
be responsible for refreshing it's own state)
- Integrating the schema manager logic into the `Manager`. Schema state
can be baked into the rule table proto also
  - Etc
- Compilation to WASM
  - Necessary optimisations/trimming of superfluous logic/dependencies
- Separation of PDP<->ePDP functionality (traces, logging etc) via build
time constraints
  - Etc

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>