enhancement: Stickier PDP IDs
This can help reduce the noise in audit logs because the ID is "sticky".
Otherwise, the same PDP could get different IDs on each restart even
though it's still running on the same hardware. These IDs are still
perfectly opaque and don't reveal anything other than the fact that the
PDP is running on the same anonymous machine every time it's restarted.

Signed-off-by: Charith Ellawala <charith@cerbos.dev>