fix: Fix e2e overlay test
The overlay cerbos instance and cerbos admin instance for the overlay
e2e tests were running in seperate pods (talking to the same postgres
instance). Storage events propogate in process, so the cerbos instance
used for Check requests did not see the updated policy state. This
coincidentally worked before because policies were loaded lazily. In the
new pre-compiled rule table world, this is not the case. Adding a
refresh after the AdminSuite run ensures policy updates were reflected
in the other cerbos process.

Also increased a `sleep` in the blob e2e test to give more time for
events to propogate to the rule table correctly.

Signed-off-by: Sam Lock <sam@swlock.co.uk>