fix: Fix e2e blob test by guaranteeing policy existence
Previously, there was a race condition between policies being added to
the bucket (and reflected in the rule table via event propagation) and
the tests running themselves. This fixes that by ensuring the minio
bucket houses the policies prior to the cerbos pod starting, so that the
rule table will be initialised with the correct state.

Signed-off-by: Sam Lock <sam@swlock.co.uk>