chore: Use OIDC to authenticate to publish npm packages (#2734)
I have configured GitHub Actions as a [trusted
publisher](https://docs.npmjs.com/trusted-publishers) for the `cerbos`
and `cerbosctl` packages (and the native dependencies e.g.
`@cerbos/cerbos-darwin-arm64`).

This PR removes the authentication token so that we use OIDC to
authenticate when publishing. Unfortunately npm only allows one workflow
to be connected so I have removed the manual override but hopefully that
won't be necessary going forward.

Signed-off-by: Andrew Haines <haines@cerbos.dev>