chore: Better node dedup in query planner (#2735)
We hit ~a bug~ an inefficiency where Go map iteration was causing
inconsistent ordering with `Expr_CreateStruct` entries. This led to
outputs with many OR'd duplicate sets of conditions (namely in rules
with wildcard roles and with many roles being resolved against those
rules).

I've tried to be a bit thorough in covering a few cases, but I'll admit
that I'm shooting from the hip somewhat. However, this feels low risk
given the context. Worst case, we'll miss a dedup opportunity and end up
with duplicate nodes (logical no-ops), which is the situation we were in
before.

---------

Signed-off-by: Sam Lock <sam@swlock.co.uk>