chore(ci): Changelog management
We used to generate the changelog from git commits based on conventional
commit information but lately that has caused a lot of confusion and
despair. This change introduces a new system where the user-facing
changelog entries are maintained separately.

Add a new changelog entry with:

```console
just changelog-entry TYPE DESCRIPTION
```

The `TYPE` could be any of the following:

- breaking
- chore
- deprecation
- docs
- enhancement
- feature
- fix
- highlight

Adding a changelog entry with each PR is highly recommended but we won't
enforce it mechanically because it doesn't make sense to write one for
each and every PR (which is what led us here in the first place).

Signed-off-by: Charith Ellawala <charith@cerbos.dev>