chore: Changelog management (#2804)
We used to generate the changelog from git commits based on conventional
commit information but lately that has caused a lot of confusion and
despair because some changes span multiple commits. This change
introduces a new system where the user-facing changelog entries are
maintained separately.

Add a new changelog entry with:

```console
just changelog-entry TYPE DESCRIPTION
```

This creates a file in `.changelog` which can be edited to add as much
detail as necessary in Asciidoc format. During a release, the contents
of all files added after the previous release will be compiled into the
changelog for the new release. This is inspired by
https://github.com/hashicorp/go-changelog. I was going to use it
initially but it required a few tweaks to match our workflow and I
eventually ended up creating a mini clone tailored to this repo.

Adding a changelog entry with each PR is highly recommended but we won't
enforce it mechanically because it doesn't make sense to write one for
each and every PR (which is what led us here in the first place).

---------

Signed-off-by: Charith Ellawala <charith@cerbos.dev>