enhancement: Simplify plan with all operation (#2817)
In case a collection is a known value, then `collection.all(t,<bodyExpr>)` can be simplified as a logical AND of the
body expr evaluated for each value of `t` in the collection. The collection can be either a map or a list. Both single- and two-var operations are supported.

---------

Signed-off-by: Dennis Buduev <dbuduev@users.noreply.github.com>
Co-authored-by: Dennis Buduev <dbuduev@users.noreply.github.com>